@@ -8,7 +8,7 @@ maintainer_email = stonesoup@dstl.gov.uk
 url = https://github.com/dstl/Stone-Soup
 license = MIT
 classifiers =
-    Development Status :: 4 - Beta
+    Development Status :: 5 - Production/Stable
     Intended Audience :: Science/Research
     License :: OSI Approved :: MIT License
     Operating System :: OS Independent
@@ -16,7 +16,7 @@ classifiers =
     Topic :: Scientific/Engineering
 
 [options]
-python_requires = >=3.7
+python_requires = >=3.8
 packages = find:
 install_requires =
     matplotlib
@@ -30,17 +30,19 @@ install_requires =
 
 [options.extras_require]
 dev =
+    confluent-kafka
     flake8<5
     folium
     h5py
     pandas
+    ortools
     pillow
     plotly
     pytest-flake8
     pytest-cov
     pytest-remotedata
     Sphinx
-    sphinx_rtd_theme
+    sphinx_rtd_theme>=1.2
     sphinx-gallery>=0.10.1
 video =
     ffmpeg-python
@@ -51,8 +53,10 @@ tensorflow =
 tensornets =
     tensorflow>=2.2.0
     tensornets
-orbital = 
+orbital =
     astropy
+mfa =
+    ortools
 
 [options.packages.find]
 exclude =
