@@ -67,7 +67,7 @@
 # model.
 import numpy as np
 from datetime import datetime, timedelta
-start_time = datetime.now()
+start_time = datetime.now().replace(microsecond=0)
 
 # %%
 
@@ -83,18 +83,20 @@ from stonesoup.models.transition.linear import CombinedLinearGaussianTransitionM
 from stonesoup.types.groundtruth import GroundTruthPath, GroundTruthState
 transition_model = CombinedLinearGaussianTransitionModel([ConstantVelocity(0.05),
                                                           ConstantVelocity(0.05)])
+timesteps = [start_time]
 truth = GroundTruthPath([GroundTruthState([0, 1, 0, 1], timestamp=start_time)])
 
 for k in range(1, 21):
+    timesteps.append(start_time+timedelta(seconds=k))
     truth.append(GroundTruthState(
         transition_model.function(truth[k-1], noise=True, time_interval=timedelta(seconds=1)),
-        timestamp=start_time+timedelta(seconds=k)))
+        timestamp=timesteps[k]))
 
 # %%
 # Plot this
 
-from stonesoup.plotter import Plotterly
-plotter = Plotterly()
+from stonesoup.plotter import AnimatedPlotterly
+plotter = AnimatedPlotterly(timesteps, tail_length=0.3)
 plotter.plot_ground_truths(truth, [0, 2])
 plotter.fig
 
@@ -216,4 +218,4 @@ plotter.fig
 # .. [#] Anderson & Moore 2012, Optimal filtering,
 #        (http://users.cecs.anu.edu.au/~john/papers/BOOK/B02.PDF)
 
-# sphinx_gallery_thumbnail_number = 3
+# sphinx_gallery_thumbnail_path = '_static/sphinx_gallery/Tutorial_2.PNG'
