@@ -102,7 +102,7 @@ import random
 from ordered_set import OrderedSet
 from datetime import datetime, timedelta
 
-start_time = datetime.now()
+start_time = datetime.now().replace(microsecond=0)
 
 from stonesoup.models.transition.linear import CombinedLinearGaussianTransitionModel, ConstantVelocity
 from stonesoup.types.groundtruth import GroundTruthPath, GroundTruthState
@@ -132,19 +132,20 @@ yps = range(0, 100, 10)  # y value for prior state
 truths = OrderedSet()
 ntruths = 3  # number of ground truths in simulation
 time_max = 50  # timestamps the simulation is observed over
+timesteps = [start_time + timedelta(seconds=k) for k in range(time_max)]
 
 xdirection = 1
 ydirection = 1
 
 # Generate ground truths
 for j in range(0, ntruths):
-    truth = GroundTruthPath([GroundTruthState([0, xdirection, yps[j], ydirection], timestamp=start_time)],
-                            id=f"id{j}")
+    truth = GroundTruthPath([GroundTruthState([0, xdirection, yps[j], ydirection],
+                                              timestamp=timesteps[0])], id=f"id{j}")
 
     for k in range(1, time_max):
         truth.append(
             GroundTruthState(transition_model.function(truth[k - 1], noise=True, time_interval=timedelta(seconds=1)),
-                             timestamp=start_time + timedelta(seconds=k)))
+                             timestamp=timesteps[k]))
     truths.add(truth)
 
     # alternate directions when initiating tracks
@@ -153,11 +154,11 @@ for j in range(0, ntruths):
         ydirection *= -1
 
 # %%
-# Plot the ground truths. This is done using the :class:`~.Plotterly` class from Stone Soup.
+# Plot the ground truths. This is done using the :class:`~.AnimatedPlotterly` class from Stone Soup.
 
-from stonesoup.plotter import Plotterly
+from stonesoup.plotter import AnimatedPlotterly
 
-plotter = Plotterly()
+plotter = AnimatedPlotterly(timesteps, tail_length=1)
 plotter.plot_ground_truths(truths, [0, 2])
 plotter.fig
 
@@ -339,11 +340,9 @@ data_associator = GNNWith2DAssignment(hypothesiser)
 # Here the chosen target for observation is selected randomly using the method :meth:`choose_actions()` from the class
 # :class:`~.RandomSensorManager`.
 
-# Generate list of timesteps from ground truth timestamps
-timesteps = []
-for state in truths[0]:
-    timesteps.append(state.timestamp)
+import copy
 
+sensor_history_A = dict()
 for timestep in timesteps[1:]:
 
     # Generate chosen configuration
@@ -358,6 +357,9 @@ for timestep in timesteps[1:]:
             sensor.add_actions(actions)
 
     sensorA.act(timestep)
+
+    # Store sensor history for plotting
+    sensor_history_A[timestep] = copy.copy(sensorA)
 
     # Observe this ground truth
     # i.e. {z}k
@@ -375,12 +377,56 @@ for timestep in timesteps[1:]:
             track.append(hypothesis.prediction)
 
 # %%
-# Plot ground truths, tracks and uncertainty ellipses for each target.
+# Plot ground truths, tracks and uncertainty ellipses for each target. This uses the Stone Soup
+# :class:`~.AnimatedPlotterly`, with added code to plot the field of view of the sensor.
 
-plotterA = Plotterly()
+import plotly.graph_objects as go
+from stonesoup.functions import pol2cart
+
+plotterA = AnimatedPlotterly(timesteps, tail_length=1, sim_duration=10)
 plotterA.plot_sensors(sensorA)
 plotterA.plot_ground_truths(truths, [0, 2])
-plotterA.plot_tracks(tracksA, [0, 2], uncertainty=True)
+plotterA.plot_tracks(tracksA, [0, 2], uncertainty=True, plot_history=False)
+
+
+def plot_sensor_fov(fig, sensor_history):
+    # Plot sensor field of view
+    trace_base = len(fig.data)
+    fig.add_trace(go.Scatter(mode='lines',
+                             line=go.scatter.Line(color='black',
+                                                  dash='dash')))
+
+    for frame in fig.frames:
+        traces_ = list(frame.traces)
+        data_ = list(frame.data)
+        x = [0, 0]
+        y = [0, 0]
+        timestring = frame.name
+        timestamp = datetime.strptime(timestring, "%Y-%m-%d %H:%M:%S")
+
+        if timestamp in sensor_history:
+            sensor = sensor_history[timestamp]
+            for i, fov_side in enumerate((-1, 1)):
+                range_ = min(getattr(sensor, 'max_range', np.inf), 100)
+                x[i], y[i] = pol2cart(range_,
+                                      sensor.dwell_centre[0, 0]
+                                      + sensor.fov_angle / 2 * fov_side) \
+                             + sensor.position[[0, 1], 0]
+        else:
+            continue
+
+        data_.append(go.Scatter(x=[x[0], sensor.position[0], x[1]],
+                                y=[y[0], sensor.position[1], y[1]],
+                                mode="lines",
+                                line=go.scatter.Line(color='black',
+                                                     dash='dash'),
+                                showlegend=False))
+        traces_.append(trace_base)
+        frame.traces = traces_
+        frame.data = data_
+
+
+plot_sensor_fov(plotterA.fig, sensor_history_A)
 plotterA.fig
 
 # %%
@@ -409,6 +455,7 @@ plotterA.fig
 # The chosen action is given to the sensor, measurements are made and the tracks updated based on these measurements.
 # Predictions are made for tracks which have not been observed by the sensor.
 
+sensor_history_B = dict()
 for timestep in timesteps[1:]:
 
     # Generate chosen configuration
@@ -423,6 +470,9 @@ for timestep in timesteps[1:]:
             sensor.add_actions(actions)
 
     sensorB.act(timestep)
+
+    # Store sensor history for plotting
+    sensor_history_B[timestep] = copy.copy(sensorB)
 
     # Observe this ground truth
     # i.e. {z}k
@@ -441,10 +491,11 @@ for timestep in timesteps[1:]:
 # %%
 # Plot ground truths, tracks and uncertainty ellipses for each target.
 
-plotterB = Plotterly()
+plotterB = AnimatedPlotterly(timesteps, tail_length=1, sim_duration=10)
 plotterB.plot_sensors(sensorB)
 plotterB.plot_ground_truths(truths, [0, 2])
-plotterB.plot_tracks(tracksB, [0, 2], uncertainty=True)
+plotterB.plot_tracks(tracksB, [0, 2], uncertainty=True, plot_history=False)
+plot_sensor_fov(plotterB.fig, sensor_history_B)
 plotterB.fig
 
 # %%
@@ -459,12 +510,29 @@ plotterB.fig
 # and SIAP metrics can be found in the Metrics Example.
 
 from stonesoup.metricgenerator.ospametric import OSPAMetric
-ospa_generator = OSPAMetric(c=40, p=1)
+ospa_generatorA = OSPAMetric(c=40, p=1,
+                             generator_name='RandomSensorManager',
+                             tracks_key='tracksA',
+                             truths_key='truths')
+
+ospa_generatorB = OSPAMetric(c=40, p=1,
+                             generator_name='BruteForceSensorManager',
+                             tracks_key='tracksB',
+                             truths_key='truths')
 
 from stonesoup.metricgenerator.tracktotruthmetrics import SIAPMetrics
 from stonesoup.measures import Euclidean
-siap_generator = SIAPMetrics(position_measure=Euclidean((0, 2)),
-                             velocity_measure=Euclidean((1, 3)))
+siap_generatorA = SIAPMetrics(position_measure=Euclidean((0, 2)),
+                              velocity_measure=Euclidean((1, 3)),
+                              generator_name='RandomSensorManager',
+                              tracks_key='tracksA',
+                              truths_key='truths')
+
+siap_generatorB = SIAPMetrics(position_measure=Euclidean((0, 2)),
+                              velocity_measure=Euclidean((1, 3)),
+                              generator_name='BruteForceSensorManager',
+                              tracks_key='tracksB',
+                              truths_key='truths')
 
 # %%
 # The SIAP metrics require an associator to associate tracks to ground truths. This is done using the
@@ -482,33 +550,36 @@ associator = TrackToTruth(association_threshold=30)
 # each time step. This gives a representation of the overall uncertainty of the tracking over time.
 
 from stonesoup.metricgenerator.uncertaintymetric import SumofCovarianceNormsMetric
-uncertainty_generator = SumofCovarianceNormsMetric()
+uncertainty_generatorA = SumofCovarianceNormsMetric(generator_name='RandomSensorManager',
+                                                    tracks_key='tracksA')
+
+uncertainty_generatorB = SumofCovarianceNormsMetric(generator_name='BruteForceSensorManager',
+                                                    tracks_key='tracksB')
 
 # %%
 # A metric manager is used for the generation of metrics on multiple :class:`~.GroundTruthPath` and
 # :class:`~.Track` objects. This takes in the metric generators, as well as the associator required for the
 # SIAP metrics.
 #
-# We must use a different metric manager for each sensor management method. This is because each sensor manager
-# generates different track data which is then used in the metric manager.
+# We input the metric generators for both sensor management methods into the same metric manager.
 
-from stonesoup.metricgenerator.manager import SimpleManager
+from stonesoup.metricgenerator.manager import MultiManager
 
-metric_managerA = SimpleManager([ospa_generator, siap_generator, uncertainty_generator],
-                                associator=associator)
-
-metric_managerB = SimpleManager([ospa_generator, siap_generator, uncertainty_generator],
-                                associator=associator)
+metric_manager = MultiManager([ospa_generatorA,
+                               ospa_generatorB,
+                               siap_generatorA,
+                               siap_generatorB,
+                               uncertainty_generatorA,
+                               uncertainty_generatorB],
+                               associator=associator)
 
 # %%
 # For each time step, data is added to the metric manager on truths and tracks. The metrics themselves can then be
 # generated from the metric manager.
 
-metric_managerA.add_data(truths, tracksA)
-metric_managerB.add_data(truths, tracksB)
+metric_manager.add_data({'truths': truths, 'tracksA': tracksA, 'tracksB': tracksB})
 
-metricsA = metric_managerA.generate_metrics()
-metricsB = metric_managerB.generate_metrics()
+metrics = metric_manager.generate_metrics()
 
 # %%
 # OSPA metric
@@ -516,22 +587,10 @@ metricsB = metric_managerB.generate_metrics()
 #
 # First we look at the OSPA metric. This is plotted over time for each sensor manager method.
 
-import matplotlib.pyplot as plt
+from stonesoup.plotter import MetricPlotter
 
-ospa_metricA = metricsA['OSPA distances']
-ospa_metricB = metricsB['OSPA distances']
-
-fig = plt.figure()
-ax = fig.add_subplot(1, 1, 1)
-ax.plot([i.timestamp for i in ospa_metricA.value],
-        [i.value for i in ospa_metricA.value],
-        label='RandomSensorManager')
-ax.plot([i.timestamp for i in ospa_metricB.value],
-        [i.value for i in ospa_metricB.value],
-        label='BruteForceSensorManager')
-ax.set_ylabel("OSPA distance")
-ax.set_xlabel("Time")
-ax.legend()
+fig = MetricPlotter()
+fig.plot_metrics(metrics, metric_names=['OSPA distances'])
 
 # %%
 # The :class:`~.BruteForceSensorManager` generally results in a smaller OSPA distance
@@ -547,29 +606,9 @@ ax.legend()
 # deleting tracks and we have one track corresponding to each ground truth. Here we only plot positional accuracy and
 # velocity accuracy over time.
 
-fig, axes = plt.subplots(2)
-
-times = metric_managerA.list_timestamps()
-
-pa_metricA = metricsA['SIAP Position Accuracy at times']
-va_metricA = metricsA['SIAP Velocity Accuracy at times']
-
-pa_metricB = metricsB['SIAP Position Accuracy at times']
-va_metricB = metricsB['SIAP Velocity Accuracy at times']
-
-axes[0].set(title='Positional Accuracy', xlabel='Time', ylabel='PA')
-axes[0].plot(times, [metric.value for metric in pa_metricA.value],
-             label='RandomSensorManager')
-axes[0].plot(times, [metric.value for metric in pa_metricB.value],
-             label='BruteForceSensorManager')
-axes[0].legend()
-
-axes[1].set(title='Velocity Accuracy', xlabel='Time', ylabel='VA')
-axes[1].plot(times, [metric.value for metric in va_metricA.value],
-             label='RandomSensorManager')
-axes[1].plot(times, [metric.value for metric in va_metricB.value],
-             label='BruteForceSensorManager')
-axes[1].legend()
+fig2 = MetricPlotter()
+fig2.plot_metrics(metrics, metric_names=['SIAP Position Accuracy at times',
+                                         'SIAP Velocity Accuracy at times'])
 
 # %%
 # Similar to the OSPA distances the :class:`~.BruteForceSensorManager`
@@ -582,20 +621,8 @@ axes[1].legend()
 # Finally we look at the uncertainty metric which computes the sum of covariance matrix norms of each state at each
 # time step. This is plotted over time for each sensor manager method.
 
-uncertainty_metricA = metricsA['Sum of Covariance Norms Metric']
-uncertainty_metricB = metricsB['Sum of Covariance Norms Metric']
-
-fig = plt.figure()
-ax = fig.add_subplot(1, 1, 1)
-ax.plot([i.timestamp for i in uncertainty_metricA.value],
-        [i.value for i in uncertainty_metricA.value],
-        label='RandomSensorManager')
-ax.plot([i.timestamp for i in uncertainty_metricB.value],
-        [i.value for i in uncertainty_metricB.value],
-        label='BruteForceSensorManager')
-ax.set_ylabel("Sum of covariance matrix norms")
-ax.set_xlabel("Time")
-ax.legend()
+fig3 = MetricPlotter()
+fig3.plot_metrics(metrics, metric_names=['Sum of Covariance Norms Metric'])
 
 # sphinx_gallery_thumbnail_number = 6
 
