@@ -53,7 +53,7 @@ class Associator(Base):
     """Associator base class
 
     An associator is used to associate objects for the generation of
-    metrics. It returns a :class:`~.AssociationSet` containing
+    metrics. It returns an :class:`~.AssociationSet` containing
     a set of :class:`~.Association`
     objects.
     """
