@@ -23,6 +23,13 @@ from ..types.update import Update
 from ..updater import Updater
 
 
+from .neighbour import NearestNeighbour
+from .neighbour import GlobalNearestNeighbour
+from .neighbour import GNNWith2DAssignment
+from .probability import PDA
+from .probability import JPDA
+
+
 class DetectionKDTreeMixIn(Base):
     """Detection kd-tree based mixin
 
@@ -151,6 +158,9 @@ class TPRTreeMixIn(Base):
         if self.vel_mapping is None:
             self.vel_mapping = [i + 1 for i in self.pos_mapping]
 
+        self._reset_tree()
+
+    def _reset_tree(self):
         # Create tree
         tree_property = rtree.index.Property(
             type=rtree.index.RT_TPRTree,
@@ -229,9 +239,63 @@ class TPRTreeMixIn(Base):
                 (*state_meas.ravel(), *state_meas.ravel()),
                 (0, 0)*len(self.pos_mapping),
                 (det_time, det_time + 1e-3)))
-            for track in intersected_tracks:
-                track_detections[track].add(detection)
+            try:
+                for track in intersected_tracks:
+                    track_detections[track].add(detection)
+            except KeyError:
+                self._reset_tree()
+                return self.generate_hypotheses(tracks, detections, timestamp, **kwargs)
 
         return {track: self.hypothesiser.hypothesise(
             track, track_detections[track], timestamp, **kwargs)
             for track in tracks}
+
+
+class DetectionKDTreeNN(DetectionKDTreeMixIn, NearestNeighbour):
+    '''DetectionKDTreeNN from NearestNeighbour and DetectionKDTreeMixIn'''
+    pass
+
+
+class DetectionKDTreeGNN(DetectionKDTreeMixIn, GlobalNearestNeighbour):
+    '''DetectionKDTreeGNN from GlobalNearestNeighbour and DetectionKDTreeMixIn'''
+    pass
+
+
+class DetectionKDTreeGNN2D(DetectionKDTreeMixIn, GNNWith2DAssignment):
+    '''DetectionKDTreeGNN2D from GNNWith2DAssignment and DetectionKDTreeMixIn'''
+    pass
+
+
+class TPRTreeNN(TPRTreeMixIn, NearestNeighbour):
+    '''TPRTreeNN from NearestNeighbour and TPRTreeMixIn'''
+    pass
+
+
+class TPRTreeGNN(TPRTreeMixIn, GlobalNearestNeighbour):
+    '''TPRTreeGNN from GlobalNearestNeighbour and TPRTreeMixIn'''
+    pass
+
+
+class TPRTreeGNN2D(TPRTreeMixIn, GNNWith2DAssignment):
+    '''TPRTreeGNN2D from GNNWith2DAssignment and TPRTreeMixIn'''
+    pass
+
+
+class KDTreePDA(DetectionKDTreeMixIn, PDA):
+    '''KDTreePDA from PDA and DetectionKDTreeMixin'''
+    pass
+
+
+class KDTreeJPDA(DetectionKDTreeMixIn, JPDA):
+    '''KDTreeJPDA from JPDA and DetectionKDTreeMixin'''
+    pass
+
+
+class TPRTreePDA(TPRTreeMixIn, PDA):
+    '''TPRTreePDA from PDA and TPRTreeMixIn'''
+    pass
+
+
+class TPRTreeJPDA(TPRTreeMixIn, JPDA):
+    '''TPRTreeJPDA from JPDA and TPRTreeMixIn'''
+    pass
