@@ -6,9 +6,13 @@ from ..base import Base
 class MetricGenerator(Base):
     """Metric Generator base class
 
-    Generates :class:`~.Metric` objects used to asses the performance of a
+    Generates :class:`~.Metric` objects used to assess the performance of a
     tracker using data held in a :class:`~.MetricManager` object
     """
+
+    @staticmethod
+    def _get_data(manager, key):
+        return manager.states_sets[key]
 
     @abstractmethod
     def compute_metric(self, manager, **kwargs):
