@@ -9,7 +9,7 @@ from ..plotter import Plotter
 
 
 class TwoDPlotter(PlotGenerator):
-    """:class:`~.MetricManager` for the plotting data
+    """:class:`~.MetricGenerator` for the plotting data
 
     Plots of :class:`~.Track`, :class:`~.Detection` and
     :class:`~.GroundTruthPath` objects in two dimensions.
@@ -24,6 +24,18 @@ class TwoDPlotter(PlotGenerator):
                                  doc='If True the plot includes uncertainty ellipses')
     particle: bool = Property(default=False,
                               doc='If True the plot includes particles')
+    tracks_key: str = Property(doc='Key to access set of tracks added to MetricManager',
+                               default='tracks')
+    truths_key: str = Property(doc="Key to access set of ground truths added to MetricManager. "
+                                   "Or key to access a second set of tracks for track-to-track "
+                                   "metric generation",
+                               default='groundtruth_paths')
+    detections_key: str = Property(doc="Key to access desired set of detections added "
+                                       "to MetricManager",
+                                   default='detections')
+    generator_name: str = Property(doc="Unique identifier to use when accessing generated "
+                                       "plots from MultiManager",
+                                   default='tracker_plot')
 
     def compute_metric(self, manager, *args, **kwargs):
         """Compute the metric using the data in the metric manager
@@ -39,9 +51,16 @@ class TwoDPlotter(PlotGenerator):
             Contains a matplotlib figure
         """
 
-        metric = self.plot_tracks_truth_detections(manager.tracks,
-                                                   manager.groundtruth_paths,
-                                                   manager.detections,
+        if self.truths_key in manager.states_sets.keys():
+            groundtruth_paths = self._get_data(manager, self.truths_key)
+        if self.tracks_key in manager.states_sets.keys():
+            tracks = self._get_data(manager, self.tracks_key)
+        if self.detections_key in manager.states_sets.keys():
+            detections = self._get_data(manager, self.detections_key)
+
+        metric = self.plot_tracks_truth_detections(tracks,
+                                                   groundtruth_paths,
+                                                   detections,
                                                    self.uncertainty,
                                                    self.particle)
         return metric
@@ -52,7 +71,7 @@ class TwoDPlotter(PlotGenerator):
 
         Parameters
         ----------
-        tracks: set of :class:`~.Track`
+        tracks: list of set of :class:`~.Track`
             Objects to be plotted as tracks
         groundtruth_paths: set of :class:`~.GroundTruthPath`
             Objects to be plotted as truths
@@ -64,25 +83,30 @@ class TwoDPlotter(PlotGenerator):
             If True, function plots particles.
 
         Returns
-        ----------
+        -------
         TimeRangePlottingMetric
             Contains the produced plot
         """
 
         plotter = Plotter()  # initialises axes using Plotter class
 
-        plotter.plot_measurements(detections, [self.detection_indices[0],
-                                               self.detection_indices[1]],
-                                  color='tab:blue')
+        plotter.ax.set_title(self.generator_name)
 
-        if groundtruth_paths:
+        if detections is not None:
+            plotter.plot_measurements(detections, [self.detection_indices[0],
+                                                   self.detection_indices[1]],
+                                      color='tab:blue')
+        else:
+            detections = []
+
+        if groundtruth_paths is not None:
             plotter.plot_ground_truths(groundtruth_paths, [self.gtruth_indices[0],
                                                            self.gtruth_indices[1]],
                                        linestyle=':')
         else:
             groundtruth_paths = []
 
-        if tracks:
+        if tracks is not None:
             plotting_tracks = set()
             for track in tracks:
                 if len([state for state in track.states if not isinstance(
@@ -95,16 +119,19 @@ class TwoDPlotter(PlotGenerator):
             if uncertainty:
                 plotter.plot_tracks(plotting_tracks, [self.track_indices[0],
                                                       self.track_indices[1]],
-                                    uncertainty=True)
+                                    uncertainty=True,
+                                    track_label=self.tracks_key)
 
             elif particle:
                 plotter.plot_tracks(plotting_tracks, [self.track_indices[0],
                                                       self.track_indices[1]],
-                                    particle=True)
+                                    particle=True,
+                                    track_label=self.tracks_key)
 
             else:
                 plotter.plot_tracks(plotting_tracks, [self.track_indices[0],
-                                                      self.track_indices[1]])
+                                                      self.track_indices[1]],
+                                    track_label=self.tracks_key)
         else:
             tracks = []
 
