@@ -3,7 +3,7 @@
 import numpy as np
 import pytest
 
-from ..manager import SimpleManager
+from ..manager import MultiManager
 from ..pcrbmetric import PCRBMetric
 from ...types.groundtruth import GroundTruthState
 from ...types.array import StateVector, StateVectors, CovarianceMatrix
@@ -129,12 +129,17 @@ def test_compute_pcrb_single(prior, transition_model, measurement_model, groundt
 )
 def test_computemetric(prior, transition_model, measurement_model, groundtruth,
                        sensor_locations, irf_overall, position_mapping, velocity_mapping):
-    pcrb = PCRBMetric(prior, transition_model, measurement_model, sensor_locations,
-                      position_mapping, velocity_mapping, irf_overall)
+    pcrb = PCRBMetric(prior=prior,
+                      transition_model=transition_model,
+                      measurement_model=measurement_model,
+                      sensor_locations=sensor_locations,
+                      position_mapping=position_mapping,
+                      velocity_mapping=velocity_mapping,
+                      irf=irf_overall)
 
-    manager = SimpleManager([pcrb])
+    manager = MultiManager([pcrb])
 
-    manager.add_data({groundtruth})
+    manager.add_data({'groundtruth_paths': groundtruth})
 
     metric = pcrb.compute_metric(manager)[0]
 
@@ -163,7 +168,7 @@ def test_computemetric(prior, transition_model, measurement_model, groundtruth,
     # NOTE: Checking value of inverse_j is not necessary since rmse values are derived from
     # inverse_j, therefore if they are correct, then inverse_j is also correct.
     assert 'inverse_j' in metric.value
-    assert metric.value['track'] == groundtruth
+    assert metric.value['track'].states == groundtruth.states
     assert np.allclose(metric.value['position_RMSE'], expected_pos_rmse)
     if velocity_mapping is not None:
         assert np.allclose(metric.value['velocity_RMSE'], expected_vel_rmse)
