@@ -3,7 +3,6 @@ from abc import ABC, abstractmethod
 from datetime import datetime, timedelta
 from itertools import chain
 from typing import Collection, Iterable, Union, List, Optional, Tuple, Dict
-
 import numpy as np
 from matplotlib import animation as animation
 from matplotlib import pyplot as plt
@@ -13,7 +12,10 @@ from matplotlib.patches import Ellipse
 from scipy.integrate import quad
 from scipy.optimize import brentq
 from scipy.stats import kde
-
+try:
+    from plotly import colors
+except ImportError:
+    colors = None
 try:
     import plotly.graph_objects as go
 except ImportError:
@@ -29,18 +31,18 @@ from .base import Base, Property
 
 from .models.base import LinearModel, Model
 
-from enum import Enum
+from enum import IntEnum
 
 
-class Dimension(Enum):
+class Dimension(IntEnum):
     """Dimension Enum class for specifying plotting parameters in the Plotter class.
     Used to sanitize inputs for the dimension attribute of Plotter().
 
     Attributes
     ----------
-    TWO: str
+    TWO: int
         Specifies 2D plotting for Plotter object
-    THREE: str
+    THREE: int
         Specifies 3D plotting for Plotter object
     """
     TWO = 2  # 2D plotting mode (original plotter.py functionality)
@@ -64,10 +66,11 @@ class _Plotter(ABC):
         raise NotImplementedError
 
     @abstractmethod
-    def plot_sensors(self, sensors, sensor_label="Sensors", **kwargs):
+    def plot_sensors(self, sensors, mapping, sensor_label="Sensors", **kwargs):
         raise NotImplementedError
 
-    def _conv_measurements(self, measurements, mapping, measurement_model=None) -> \
+    def _conv_measurements(self, measurements, mapping, measurement_model=None,
+                           convert_measurements=True) -> \
             Tuple[Dict[detection.Detection, StateVector], Dict[detection.Clutter, StateVector]]:
         conv_detections = {}
         conv_clutter = {}
@@ -76,11 +79,12 @@ class _Plotter(ABC):
             if meas_model is None:
                 meas_model = measurement_model  # measurement_model from input
 
-            if isinstance(meas_model, LinearModel):
+            if not convert_measurements:
+                state_vec = state.state_vector[mapping, :]
+            elif isinstance(meas_model, LinearModel):
                 model_matrix = meas_model.matrix()
                 inv_model_matrix = np.linalg.pinv(model_matrix)
                 state_vec = (inv_model_matrix @ state.state_vector)[mapping, :]
-
             elif isinstance(meas_model, Model):
                 try:
                     state_vec = meas_model.inverse_function(state)[mapping, :]
@@ -102,7 +106,6 @@ class _Plotter(ABC):
             else:
                 warnings.warn(f'Unknown type {type(state)}')
                 continue
-
         return conv_detections, conv_clutter
 
 
@@ -118,6 +121,8 @@ class Plotter(_Plotter):
     ----------
     dimension: enum \'Dimension\'
         Optional parameter to specify 2D or 3D plotting. Default is 2D plotting.
+    plot_timeseries: bool
+        Specify whether data to be plotted is time series data. Default False
     \\*\\*kwargs: dict
         Additional arguments to be passed to plot function. For example, figsize (Default is
         (10, 6)).
@@ -138,6 +143,8 @@ class Plotter(_Plotter):
         figure_kwargs.update(kwargs)
         if isinstance(dimension, type(Dimension.TWO)):
             self.dimension = dimension
+        elif isinstance(dimension, int):
+            self.dimension = Dimension(dimension)
         else:
             raise TypeError("%s is an unsupported type for \'dimension\'; "
                             "expected type %s" % (type(dimension), type(Dimension.TWO)))
@@ -175,6 +182,8 @@ class Plotter(_Plotter):
             for iteration.
         mapping: list
             List of items specifying the mapping of the position components of the state space.
+        truths_label: str
+            Label for truth data. Default is "Ground Truth"
         \\*\\*kwargs: dict
             Additional arguments to be passed to plot function. Default is ``linestyle="--"``.
 
@@ -212,7 +221,7 @@ class Plotter(_Plotter):
         return artists
 
     def plot_measurements(self, measurements, mapping, measurement_model=None,
-                          measurements_label="Measurements", **kwargs):
+                          measurements_label="Measurements", convert_measurements=True, **kwargs):
         """Plots measurements
 
         Plots detections and clutter, generating a legend automatically. Detections are plotted as
@@ -231,6 +240,11 @@ class Plotter(_Plotter):
         measurement_model : :class:`~.Model`, optional
             User-defined measurement model to be used in finding measurement state inverses if
             they cannot be found from the measurements themselves.
+        measurements_label : str
+            Label for the measurements.  Default is "Measurements".
+        convert_measurements : bool
+            Should the measurements be converted from measurement space to state space before
+            being plotted. Default is True
         \\*\\*kwargs: dict
             Additional arguments to be passed to plot function for detections. Defaults are
             ``marker='o'`` and ``color='b'``.
@@ -254,7 +268,8 @@ class Plotter(_Plotter):
 
         plot_detections, plot_clutter = self._conv_measurements(measurements_set,
                                                                 mapping,
-                                                                measurement_model)
+                                                                measurement_model,
+                                                                convert_measurements)
 
         artists = []
         if plot_detections:
@@ -282,7 +297,7 @@ class Plotter(_Plotter):
         return artists
 
     def plot_tracks(self, tracks, mapping, uncertainty=False, particle=False, track_label="Tracks",
-                    err_freq=1, **kwargs):
+                    err_freq=1, same_color=False, **kwargs):
         """Plots track(s)
 
         Plots each track generated, generating a legend automatically. If ``uncertainty=True``
@@ -312,6 +327,8 @@ class Plotter(_Plotter):
         err_freq: int
             Frequency of error bar plotting on tracks. Default value is 1, meaning
             error bars are plotted at every track step.
+        same_color: bool
+            Should all the tracks have the same color. Default False
         \\*\\*kwargs: dict
             Additional arguments to be passed to plot function. Defaults are ``linestyle="-"``,
             ``marker='s'`` for :class:`~.Update` and ``marker='o'`` for other states.
@@ -357,6 +374,8 @@ class Plotter(_Plotter):
                     linestyle='',
                     color=plt.getp(line[0], 'color')))
             track_colors[track] = plt.getp(line[0], 'color')
+            if same_color:
+                tracks_kwargs['color'] = plt.getp(line[0], 'color')
 
         if tracks:  # If no tracks `line` won't be defined
             # Assuming a single track or all plotted as the same colour then the following will
@@ -453,20 +472,23 @@ class Plotter(_Plotter):
 
         return artists
 
-    def plot_sensors(self, sensors, sensor_label="Sensors", **kwargs):
+    def plot_sensors(self, sensors, mapping=None, sensor_label="Sensors", **kwargs):
         """Plots sensor(s)
 
-        Plots sensors.  Users can change the color and marker of detections using keyword
+        Plots sensors.  Users can change the color and marker of sensors using keyword
         arguments. Default is a black 'x' marker.
 
         Parameters
         ----------
         sensors : Collection of :class:`~.Sensor`
             Sensors to plot
+        mapping: list
+            List of items specifying the mapping of the position components of the
+            sensor's position. Default is either [0, 1] or [0, 1, 2] depending on `self.dimension`
         sensor_label: str
-            Label to apply to all tracks for legend.
+            Label to apply to all sensors for legend.
         \\*\\*kwargs: dict
-            Additional arguments to be passed to plot function for detections. Defaults are
+            Additional arguments to be passed to plot function for sensors. Defaults are
             ``marker='x'`` and ``color='black'``.
 
         Returns
@@ -481,16 +503,19 @@ class Plotter(_Plotter):
         if not isinstance(sensors, Collection):
             sensors = {sensors}  # Make a set of length 1
 
+        if mapping is None:
+            mapping = list(range(self.dimension))
+
         artists = []
         for sensor in sensors:
             if self.dimension is Dimension.TWO:  # plots the sensors in xy
-                artists.append(self.ax.scatter(sensor.position[0],
-                                               sensor.position[1],
+                artists.append(self.ax.scatter(sensor.position[mapping[0]],
+                                               sensor.position[mapping[1]],
                                                **sensor_kwargs))
             elif self.dimension is Dimension.THREE:  # plots the sensors in xyz
-                artists.extend(self.ax.plot3D(sensor.position[0],
-                                              sensor.position[1],
-                                              sensor.position[2],
+                artists.extend(self.ax.plot3D(sensor.position[mapping[0]],
+                                              sensor.position[mapping[1]],
+                                              sensor.position[mapping[2]],
                                               **sensor_kwargs))
             else:
                 raise NotImplementedError('Unsupported dimension type for sensor plotting')
@@ -534,14 +559,14 @@ class Plotter(_Plotter):
                            [state.state_vector[2] for state in ghosts],
                            linestyle="")
 
-    def plot_density(self, state_sequences: Iterable[StateMutableSequence],
+    def plot_density(self, state_sequences: Collection[StateMutableSequence],
                      index: Union[int, None] = -1,
                      mapping=(0, 2), n_bins=300, **kwargs):
         """
 
         Parameters
         ----------
-        state_sequences : an iterable of :class:`~.StateMutableSequence`
+        state_sequences : a collection of :class:`~.StateMutableSequence`
             Set of tracks which will be plotted. If not a set, and instead a single
             :class:`~.Track` type, the argument is modified to be a set to allow for iteration.
         index: int
@@ -626,6 +651,312 @@ class _HandlerEllipse(HandlerPatch):
         return [p]
 
 
+class MetricPlotter(ABC):
+    """Class for plotting Stone Soup metrics using matplotlib
+
+    A plotting class which is used to simplify the process of plotting metrics.
+    Legends are automatically generated with each plot.
+
+    """
+    def __init__(self):
+        self.fig = None
+        self.axes = None
+        self.plottable_metrics = ["OSPA distances",
+                                  "GOSPA Metrics",
+                                  "SIAP Completeness at times",
+                                  "SIAP Ambiguity at times",
+                                  "SIAP Spuriousness at times",
+                                  "SIAP Position Accuracy at times",
+                                  "SIAP Velocity Accuracy at times",
+                                  "SIAP ID Completeness at times",
+                                  "SIAP ID Correctness at times",
+                                  "SIAP ID Ambiguity at times",
+                                  "PCRB Metrics",
+                                  "Sum of Covariance Norms Metric",
+                                  "Mean of Covariance Norms Metric"
+                                  ]
+
+    def plot_metrics(self, metrics, generator_names=None, metric_names=None,
+                     combine_plots=True, **kwargs):
+        """Plots metrics
+
+        Plots each plottable metric passed in to :attr:`metrics` across a series of subplots
+        and generates legend(s) automatically. Metrics are plotted as lines with default colors.
+
+        Users can change linestyle, color and marker or other features using keyword arguments.
+        Any changes will apply to all metrics.
+
+        Parameters
+        ----------
+        metrics : dict of :class:`~.Metric`
+            Dictionary of generated metrics to be plotted.
+        generator_names: list of str
+            Generator(s) to extract specific metrics from :attr:`metrics` for plotting.
+            Default None to take all metrics.
+        metric_names: list of str
+            Specific metric(s) to extract from :class:`~.MetricGenerator` for plotting.
+            Default None to take all metrics in generators.
+        combine_plots: bool
+            Plot metrics of same type on the same subplot. Default True.
+        \\*\\*kwargs: dict
+            Additional arguments to be passed to plot function. Default is ``linestyle="-"``.
+
+        Returns
+        -------
+        : :class:`matplotlib.pyplot.figure`
+            Figure containing subplots displaying all plottable metrics.
+        """
+        metrics_kwargs = dict(linestyle="-")
+        metrics_kwargs.update(kwargs)
+
+        generator_names = list(metrics.keys()) if generator_names is None else generator_names
+
+        # warning for user input metrics that will not be plotted
+        if metric_names is not None:
+            for metric_name in metric_names:
+                if metric_name not in self.plottable_metrics:
+                    warnings.warn(f"{metric_name} "
+                                  f"is not a plottable metric and will not be plotted")
+        else:
+            metric_names = self.extract_metric_types(metrics)
+
+        metrics_to_plot = self._extract_plottable_metrics(metrics, generator_names, metric_names)
+
+        if combine_plots:
+            self.combine_plots(metrics_to_plot, metrics_kwargs)
+        else:
+            self.plot_separately(metrics_to_plot, metrics_kwargs)
+
+    def _extract_plottable_metrics(self, metrics, generator_names, metric_names):
+        """
+        Extract all plottable metrics from dict of generated metrics.
+
+        Parameters
+        ----------
+        metrics: dict of :class:`~.Metric`
+            Dictionary of generated metrics.
+        generator_names: list of str
+            Generator(s) to extract specific metrics from :attr:`metrics` for plotting.
+        metric_names: list of str
+            Specific metric(s) to extract from :class:`~.MetricGenerator` for plotting.
+
+        Returns
+        -------
+        : dict
+            Dict of all plottable metrics.
+        """
+        metrics_dict = dict()
+
+        for generator_name in generator_names:
+            for metric_name in metric_names:
+                if metric_name in metrics[generator_name].keys() and \
+                        metric_name in self.plottable_metrics:
+                    if generator_name not in metrics_dict.keys():
+                        metrics_dict[generator_name] = \
+                            {metric_name: metrics[generator_name][metric_name]}
+                    else:
+                        metrics_dict[generator_name][metric_name] = \
+                            metrics[generator_name][metric_name]
+
+        return metrics_dict
+
+    def _count_subplots(self, metrics_to_plot, combine_plots):
+        """
+        Calculate number of subplots needed to plot all metrics.
+
+        Parameters
+        ----------
+        metrics_to_plot: dict of :class:`~.Metric`
+            Dictionary of metrics to be plotted.
+        combine_plots: bool
+            Specifies whether same metric types should be plotted on same subplot.
+
+        Returns
+        -------
+        : int
+            Number of subplots to generate.
+        """
+        if combine_plots:
+            metric_types = self.extract_metric_types(metrics_to_plot)
+            number_of_subplots = len(metric_types)
+
+        else:
+            number_of_subplots = 0
+            for generator in metrics_to_plot.keys():
+                number_of_subplots += len(metrics_to_plot[generator])
+
+        return number_of_subplots
+
+    @staticmethod
+    def extract_metric_types(metrics):
+        """
+        Identify the different types of metric held in dict of metrics.
+
+        Parameters
+        ----------
+        metrics: dict of :class:`~.Metric`
+            Dictionary of metrics.
+
+        Returns
+        -------
+        : list
+            Sorted list of types of metric
+        """
+        metric_types = set()
+        for generator in metrics.keys():
+            for metric_key in metrics[generator].keys():
+                metric_types.add(metric_key)
+
+        metric_types = list(metric_types)
+        metric_types.sort()
+
+        return metric_types
+
+    def combine_plots(self, metrics_to_plot, metrics_kwargs):
+        """
+        Generates one subplot for each different metric type and plots metrics of the same
+        type on same subplot. Metrics are plotted over time.
+
+        Parameters
+        ----------
+        metrics_to_plot: dict of :class:`~.Metric`
+            Dictionary of metrics to plot.
+        metrics_kwargs: dict
+            Keyword arguments to be passed to plot function.
+
+        Returns
+        -------
+        : :class:`matplotlib.pyplot.figure`
+            Figure containing subplots displaying metrics.
+        """
+        # determine how many plots required - equal to number of metric types
+        number_of_subplots = self._count_subplots(metrics_to_plot, True)
+
+        # initialise each subplot
+        self.fig, axes = plt.subplots(number_of_subplots, figsize=(10, 6*number_of_subplots))
+        self.fig.subplots_adjust(hspace=0.3)
+
+        # extract data for each subplot and plot it
+        metric_types = self.extract_metric_types(metrics_to_plot)
+
+        self.axes = axes if isinstance(axes, Iterable) else [axes]
+
+        # generate colour map for lines to be plotted
+        if 'color' not in metrics_kwargs.keys():
+            colour_map = plt.cm.rainbow(np.linspace(0, 1, len(metrics_to_plot.keys())))
+        else:
+            colour_map = metrics_kwargs['color']
+            metrics_kwargs.pop('color')
+
+        for metric_type, axis in zip(list(metric_types), self.axes):
+            artists = []
+            legend_dict = {}
+
+            colour_map_copy = iter(colour_map.copy())
+
+            for generator in metrics_to_plot.keys():
+                for metric in metrics_to_plot[generator].keys():
+                    if metric == metric_type:
+                        colour = next(colour_map_copy)
+                        metric_values = metrics_to_plot[generator][metric].value
+                        artists.extend(axis.plot([_.timestamp for _ in metric_values],
+                                                 [_.value for _ in metric_values],
+                                                 color=colour,
+                                                 **metrics_kwargs))
+
+                        metric_handle = Line2D([], [], linestyle=metrics_kwargs['linestyle'],
+                                               color=colour)
+                        legend_dict[generator] = metric_handle
+
+            # Generate legend
+            artists.append(axis.legend(handles=legend_dict.values(),
+                                       labels=legend_dict.keys()))
+
+            y_label = metric_type.split(' at times')[0]
+            artists.extend(axis.set(title=metric_type.split(' at times')[0],
+                                    xlabel="Time", ylabel=y_label))
+
+    def plot_separately(self, metrics_to_plot, metrics_kwargs):
+        """
+        Generates one subplot for each different individual metric and plots metric
+        values over time.
+
+        Parameters
+        ----------
+        metrics_to_plot: dict of :class:`~.Metric`
+            Dictionary of metrics to plot.
+        metrics_kwargs: dict
+            Keyword arguments to be passed to plot function.
+
+        Returns
+        -------
+        : :class:`matplotlib.pyplot.figure`
+            Figure containing subplots displaying metrics.
+        """
+        metrics_kwargs['color'] = metrics_kwargs['color'] if \
+            'color' in metrics_kwargs.keys() else 'blue'
+
+        # determine how many plots required - equal to number of metrics within the generators
+        number_of_subplots = self._count_subplots(metrics_to_plot, False)
+
+        # initialise each plot
+        self.fig, axes = plt.subplots(number_of_subplots, figsize=(10, 6*number_of_subplots))
+        self.fig.subplots_adjust(hspace=0.3)
+
+        # extract data for each plot and plot it
+        all_metrics = {}
+        for generator in metrics_to_plot.keys():
+            for metric in list(metrics_to_plot[generator].keys()):
+                all_metrics[f'{generator}: {metric}'] = metrics_to_plot[generator][metric]
+
+        self.axes = axes if isinstance(axes, Iterable) else [axes]
+
+        for metric, axis in zip(all_metrics.keys(), self.axes):
+            y_label = str(all_metrics[metric].title).split(' at times')[0]
+            axis.set(title=str(all_metrics[metric].title), xlabel='Time', ylabel=y_label)
+            metric_values = all_metrics[metric].value
+            axis.plot([_.timestamp for _ in metric_values],
+                      [_.value for _ in metric_values],
+                      **metrics_kwargs)
+
+            # Generate legend
+            metric_handle = Line2D([], [], linestyle=metrics_kwargs['linestyle'],
+                                   color=metrics_kwargs['color'])
+            axis.legend(handles=[metric_handle],
+                        labels=[metric.split(' at times')[0]])
+
+    def set_fig_title(self, title):
+        """
+        Set title for the figure.
+
+        Parameters
+        ----------
+        title: str
+            Figure title text.
+
+        Returns
+        -------
+        Text instance of figure title.
+        """
+        self.fig.suptitle(t=title)
+
+    def set_ax_title(self, titles):
+        """
+        Set axis titles for each axis in figure.
+
+        Parameters
+        ----------
+        titles: list of str
+            List of strings for title text for each axis.
+
+        Returns
+        -------
+        Text instance of axis titles.
+        """
+        for axis, title in zip(self.axes, titles):
+            axis.set(title=title)
+
+
 class Plotterly(_Plotter):
     """Plotting class for building graphs of Stone Soup simulations using plotly
 
@@ -652,6 +983,8 @@ class Plotterly(_Plotter):
             raise RuntimeError("Usage of Plotterly plotter requires installation of `plotly`")
         if isinstance(dimension, type(Dimension.TWO)):
             self.dimension = dimension
+        elif isinstance(dimension, int):
+            self.dimension = Dimension(dimension)
         else:
             raise TypeError("%s is an unsupported type for \'dimension\'; "
                             "expected type %s" % (type(dimension), type(Dimension.TWO)))
@@ -696,6 +1029,8 @@ class Plotterly(_Plotter):
             set to allow for iteration.
         mapping: list
             List of items specifying the mapping of the position components of the state space.
+        truths_label: str
+            Label for truth data. Default is "Ground Truth"
         \\*\\*kwargs: dict
             Additional arguments to be passed to scatter function. Default is
             ``line=dict(dash="dash")``.
@@ -723,7 +1058,7 @@ class Plotterly(_Plotter):
                 **scatter_kwargs)
 
     def plot_measurements(self, measurements, mapping, measurement_model=None,
-                          measurements_label="Measurements", **kwargs):
+                          measurements_label="Measurements", convert_measurements=True, **kwargs):
         """Plots measurements
 
         Plots detections and clutter, generating a legend automatically. Detections are plotted as
@@ -744,6 +1079,9 @@ class Plotterly(_Plotter):
             they cannot be found from the measurements themselves.
         measurements_label : str
             Label for the measurements.  Default is "Measurements".
+        convert_measurements: bool
+            Should the measurements be converted from measurement space to state space before
+            being plotted. Default is True
         \\*\\*kwargs: dict
             Additional arguments to be passed to scatter function for detections. Defaults are
             ``marker=dict(color="#636EFA")``.
@@ -759,7 +1097,8 @@ class Plotterly(_Plotter):
 
         plot_detections, plot_clutter = self._conv_measurements(measurements_set,
                                                                 mapping,
-                                                                measurement_model)
+                                                                measurement_model,
+                                                                convert_measurements)
 
         if plot_detections:
             name = measurements_label + "<br>(Detections)"
@@ -772,7 +1111,7 @@ class Plotterly(_Plotter):
                 measurement_kwargs['showlegend'] = True
             else:
                 measurement_kwargs['showlegend'] = False
-            detection_array = np.array(list(plot_detections.values()))
+            detection_array = np.asfarray(list(plot_detections.values()))
             self.fig.add_scatter(
                 x=detection_array[:, 0],
                 y=detection_array[:, 1],
@@ -791,7 +1130,7 @@ class Plotterly(_Plotter):
                 measurement_kwargs['showlegend'] = True
             else:
                 measurement_kwargs['showlegend'] = False
-            clutter_array = np.array(list(plot_clutter.values()))
+            clutter_array = np.asfarray(list(plot_clutter.values()))
             self.fig.add_scatter(
                 x=clutter_array[:, 0],
                 y=clutter_array[:, 1],
@@ -799,8 +1138,30 @@ class Plotterly(_Plotter):
                 **measurement_kwargs,
             )
 
+    def get_next_color(self):
+        """
+        Find the colour of the next plot. This approach to getting colour isn't ideal, but should
+        work in most cases...
+        Returns
+        -------
+        dist : str
+            Hex string for a colour
+        """
+        # Find how many sequences have been plotted so far. The current plot has already been added
+        # to fig.data, so -1 is needed
+        figure_index = len(self.fig.data) - 1
+
+        # Get the list of colours used for plotting
+        colorway = self.fig.layout.colorway
+        max_index = len(colorway)
+
+        # Use the modulo operator to limit the colour index to limits of the colorway.
+        # If figure_index > max_index then colours will be reused
+        color_index = figure_index % max_index
+        return colorway[color_index]
+
     def plot_tracks(self, tracks, mapping, uncertainty=False, particle=False, track_label="Tracks",
-                    ellipse_points=30, **kwargs):
+                    ellipse_points=30, same_color=False, **kwargs):
         """Plots track(s)
 
         Plots each track generated, generating a legend automatically. If ``uncertainty=True``
@@ -825,6 +1186,8 @@ class Plotterly(_Plotter):
             Label to apply to all tracks for legend.
         ellipse_points: int
             Number of points for polygon approximating ellipse shape
+        same_color: bool
+            Should all the tracks have the same colour. Default False
         \\*\\*kwargs: dict
             Additional arguments to be passed to scatter function. Defaults are
             ``marker=dict(symbol='square')`` for :class:`~.Update` and
@@ -839,6 +1202,15 @@ class Plotterly(_Plotter):
         track_kwargs.update(kwargs)
         add_legend = track_kwargs['legendgroup'] not in {trace.legendgroup
                                                          for trace in self.fig.data}
+        if same_color:
+            color = track_kwargs.get('marker', {}).get('color') or \
+                    track_kwargs.get('line', {}).get('color')
+
+            # Set the colour if it hasn't already been set
+            if color is None:
+                track_kwargs['marker'] = track_kwargs.get('marker', {})
+                track_kwargs['marker']['color'] = self.get_next_color()
+
         for track in tracks:
             scatter_kwargs = track_kwargs.copy()
             scatter_kwargs['name'] = track.id
@@ -858,15 +1230,10 @@ class Plotterly(_Plotter):
                 y=[getattr(state, 'mean', state.state_vector)[mapping[1]] for state in track],
                 text=[self._format_state_text(state) for state in track],
                 **scatter_kwargs)
-            color = self.fig.data[-1].line.color
-            if color is not None:
-                track_colors[track] = color
-            else:
-                # This approach to getting colour isn't ideal, but should work in most cases...
-                index = len(self.fig.data) - 1
-                colorway = self.fig.layout.colorway
-                max_index = len(colorway)
-                track_colors[track] = colorway[index % max_index]
+
+            track_colors[track] = (self.fig.data[-1].line.color
+                                   or self.fig.data[-1].marker.color
+                                   or self.get_next_color())
 
         if uncertainty:
             name = track_kwargs['legendgroup'] + "<br>(Ellipses)"
@@ -933,25 +1300,28 @@ class Plotterly(_Plotter):
 
         c, s = np.cos(orient), np.sin(orient)
         rotational_matrix = np.array(((c, -s), (s, c)))
-
+        points.append(points[0])
         points = np.array([[a * np.sin(i), b * np.cos(i)] for i in points])
         points = rotational_matrix @ points.T
         return points + state.mean[mapping[:2], :]
 
-    def plot_sensors(self, sensors, sensor_label="Sensors", **kwargs):
+    def plot_sensors(self, sensors, mapping=[0, 1], sensor_label="Sensors", **kwargs):
         """Plots sensor(s)
 
-        Plots sensors.  Users can change the color and marker of detections using keyword
+        Plots sensors. Users can change the color and marker of sensors using keyword
         arguments. Default is a black 'x' marker.
 
         Parameters
         ----------
         sensors : Collection of :class:`~.Sensor`
             Sensors to plot
+        mapping: list
+            List of items specifying the mapping of the position
+            components of the sensor's position.
         sensor_label: str
-            Label to apply to all tracks for legend.
+            Label to apply to all sensors for legend.
         \\*\\*kwargs: dict
-            Additional arguments to be passed to scatter function for detections. Defaults are
+            Additional arguments to be passed to scatter function for sensors. Defaults are
             ``marker=dict(symbol='x', color='black')``.
         """
 
@@ -969,8 +1339,26 @@ class Plotterly(_Plotter):
         else:
             sensor_kwargs['showlegend'] = True
 
-        sensor_xy = np.array([sensor.position[[0, 1], 0] for sensor in sensors])
+        sensor_xy = np.array([sensor.position[mapping, 0] for sensor in sensors])
         self.fig.add_scatter(x=sensor_xy[:, 0], y=sensor_xy[:, 1], **sensor_kwargs)
+
+    def hide_plot_traces(self, items_to_hide: set):
+        """Hide Plot Traces
+
+        This function allows plotting items to be invisible as default. Users can toggle the plot
+        trace to visible.
+
+        Parameters
+        ----------
+        items_to_hide : set[str]
+            The legend label (`legendgroups`) for the plot traces that should be invisible as
+            default
+        """
+        for fig_data in self.fig.data:
+            if fig_data.legendgroup in items_to_hide:
+                fig_data.visible = "legendonly"
+            else:
+                fig_data.visible = None
 
 
 class _AnimationPlotterDataClass(Base):
@@ -1046,7 +1434,7 @@ class AnimationPlotter(_Plotter):
             Additional arguments to be passed to the animation.save function
         """
         if self.animation_output is None:
-            raise ValueError("Animation hasn't been ran yet. Therefore there is no animation to "
+            raise ValueError("Animation hasn't been run yet. Therefore there is no animation to "
                              "save")
 
         self.animation_output.save(filename, **kwargs)
@@ -1070,7 +1458,7 @@ class AnimationPlotter(_Plotter):
         mapping: list
             List of items specifying the mapping of the position components of the state space.
         truths_label: str
-            Label for truth data
+            Label for truth data. Default is "Ground Truth"
         \\*\\*kwargs: dict
             Additional arguments to be passed to plot function. Default is ``linestyle="--"``.
         """
@@ -1149,7 +1537,7 @@ class AnimationPlotter(_Plotter):
             ))
 
     def plot_measurements(self, measurements, mapping, measurement_model=None,
-                          measurements_label="", **kwargs):
+                          measurements_label="", convert_measurements=True, **kwargs):
         """Plots measurements
 
         Plots detections and clutter, generating a legend automatically. Detections are plotted as
@@ -1169,7 +1557,10 @@ class AnimationPlotter(_Plotter):
             User-defined measurement model to be used in finding measurement state inverses if
             they cannot be found from the measurements themselves.
         measurements_label: str
-            Label for measurements
+            Label for measurements. Default will be "Detections" or "Clutter"
+        convert_measurements: bool
+            Should the measurements be converted from measurement space to state space before
+            being plotted. Default is True
         \\*\\*kwargs: dict
             Additional arguments to be passed to plot function for detections. Defaults are
             ``marker='o'`` and ``color='b'``.
@@ -1188,7 +1579,8 @@ class AnimationPlotter(_Plotter):
 
         plot_detections, plot_clutter = self._conv_measurements(measurements_set,
                                                                 mapping,
-                                                                measurement_model)
+                                                                measurement_model,
+                                                                convert_measurements)
 
         if measurements_label != "":
             measurements_label = measurements_label + " "
@@ -1234,11 +1626,12 @@ class AnimationPlotter(_Plotter):
         Parameters
         ----------
         times_to_plot : Iterable[datetime]
-            All the times, that the plotter should plot
+            All the times that the plotter should plot
         data : Iterable[datetime]
             All the data that should be plotted
         plot_item_expiry: timedelta
-            How long a state should be displayed for. None means the
+            How long a state should be displayed for. Default value of None
+            means data is shown indefinitely
         axis_padding: float
             How much extra space should be given around the edge of the plot
         figure_kwargs: dict
@@ -1339,8 +1732,6 @@ class AnimationPlotter(_Plotter):
             The data that will be plotted, to be plotted.
         data_list : List[List[State]]
             All the data that should be plotted
-        mapping : tuple
-            The indices of the state vector that should be plotted
         start_times : List[datetime]
             lowest (earliest) time for an item to be plotted
         end_times : List[datetime]
@@ -1368,3 +1759,836 @@ class AnimationPlotter(_Plotter):
                     lines[i].set_data([],
                                       [])
         return lines
+
+
+class AnimatedPlotterly(_Plotter):
+    """
+    Class for a 2D animated plotter that uses Plotly graph objects rather than matplotlib.
+    This gives the user the ability to see how tracking works through time, while being
+    able to interact with tracks, truths, etc, in the same way that is enabled by
+    Plotly static plots.
+
+    Simplifies the process of plotting ground truths, measurements, clutter, and tracks.
+    Tracks can be plotted with uncertainty ellipses or particles if required. Legends
+    are automatically generated with each plot.
+
+    Parameters
+    ----------
+    timesteps: Collection
+        Collection of equally-spaced timesteps. Each animation frame is a timestep.
+    tail_length: float
+        Percentage of sim time for which previous values will still be displayed for.
+        Value can be between 0 and 1. Default is 0.3.
+    equal_size: bool
+        Makes x and y axes equal when figure is resized. Default is False.
+    sim_duration: int
+        Time taken to run animation (s). Default is 6
+    \\*\\*kwargs
+        Additional arguments to be passed in the initialisation.
+
+    Attributes
+    ----------
+
+    """
+
+    def __init__(self, timesteps, tail_length=0.3, equal_size=False,
+                 sim_duration=6, **kwargs):
+        """
+        Initialise the figure and checks that inputs are correctly formatted.
+        Creates an empty frame for each timestep, and configures
+        the buttons and slider.
+
+
+        """
+        if go is None or colors is None:
+            raise RuntimeError("Usage of Plotterly plotter requires installation of `plotly`")
+
+        self.equal_size = equal_size
+
+        # checking that there are multiple timesteps
+        if len(timesteps) < 2:
+            raise ValueError("Must be at least 2 timesteps for animation.")
+
+        # checking that timesteps are evenly spaced
+        time_spaces = np.unique(np.diff(timesteps))
+
+        # gives the unique values of time gaps between timesteps. If this contains more than
+        # one value, then timesteps are not all evenly spaced which is an issue.
+        if len(time_spaces) != 1:
+            raise ValueError("Ensure timesteps are equally spaced.")
+        self.timesteps = timesteps
+
+        # checking input to tail_length
+        if tail_length > 1 or tail_length < 0:
+            raise ValueError("Tail length should be between 0 and 1")
+        self.tail_length = tail_length
+
+        # checking sim_duration
+        if sim_duration <= 0:
+            raise ValueError("Simulation duration must be positive")
+
+        # time window is calculated as sim_length * tail_length. This is
+        # the window of time for which past plots are still visible
+        self.time_window = (timesteps[-1] - timesteps[0]) * tail_length
+
+        self.colorway = colors.qualitative.Plotly[1:]  # plotting colours
+
+        self.all_masks = dict()  # dictionary to be filled up later
+
+        self.plotting_function_called = False  # keeps track if anything has been plotted or not
+        # so that only the first data plotted will override the default axis max and mins.
+
+        self.fig = go.Figure()
+
+        layout_kwargs = dict(
+            xaxis=dict(title=dict(text="<i>x</i>")),
+            yaxis=dict(title=dict(text="<i>y</i>")),
+            colorway=self.colorway,  # Needed to match colours later.
+            height=550,
+            autosize=True
+        )
+        # layout_kwargs.update(kwargs)
+        self.fig.update_layout(layout_kwargs)
+
+        # initialise frames according to simulation timesteps
+        self.fig.frames = [dict(
+            name=str(time),
+            data=[],
+            traces=[]
+        ) for time in timesteps]
+
+        self.fig.update_xaxes(range=[0, 10])
+        self.fig.update_yaxes(range=[0, 10])
+
+        frame_duration = sim_duration * 1000 / len(self.fig.frames)
+
+        # if the gap between timesteps is greater than a day, it isn't necessary
+        # to display hour and minute information, so remove this to give a cleaner display.
+        # a and b are used in the slider steps label later
+        if time_spaces[0] >= timedelta(days=1):
+            start_cut_off = None
+            end_cut_off = 10
+
+        # if the simulation is over a day long, display all information which
+        # looks clunky but is necessary
+        elif timesteps[-1] - timesteps[0] > timedelta(days=1):
+            start_cut_off = None
+            end_cut_off = None
+
+        # otherwise, remove day information and just show
+        # hours, mins, etc. which is cleaner to look at
+        else:
+            start_cut_off = 11
+            end_cut_off = None
+
+        # create button and slider
+        updatemenus = [dict(type='buttons',
+                            buttons=[{
+                                "args": [None,
+                                         {"frame": {"duration": frame_duration, "redraw": True},
+                                          "fromcurrent": True, "transition": {"duration": 0}}],
+                                "label": "Play",
+                                "method": "animate"
+                            }, {
+                                "args": [[None], {"frame": {"duration": 0, "redraw": True},
+                                                  "mode": "immediate",
+                                                  "transition": {"duration": 0}}],
+                                "label": "Stop",
+                                "method": "animate"
+                            }],
+                            direction='left',
+                            pad=dict(r=10, t=75),
+                            showactive=True, x=0.1, y=0, xanchor='right', yanchor='top')
+                       ]
+        sliders = [{'yanchor': 'top',
+                    'xanchor': 'left',
+                    'currentvalue': {'font': {'size': 16}, 'prefix': 'Time: ', 'visible': True,
+                                     'xanchor': 'right'},
+                    'transition': {'duration': frame_duration, 'easing': 'linear'},
+                    'pad': {'b': 10, 't': 50},
+                    'len': 0.9, 'x': 0.1, 'y': 0,
+                    'steps': [{'args': [[frame.name], {
+                        'frame': {'duration': 1.0, 'easing': 'linear', 'redraw': True},
+                        'transition': {'duration': 0, 'easing': 'linear'}}],
+                               'label': frame.name[start_cut_off: end_cut_off],
+                               'method': 'animate'} for frame in
+                              self.fig.frames
+                              ]}]
+        self.fig.update_layout(updatemenus=updatemenus, sliders=sliders)
+        self.fig.update_layout(kwargs)
+
+    def show(self):
+        """
+        Display the animation.
+        """
+        return self.fig
+
+    def _resize(self, data, type="track"):
+        """
+        Reshape figure so that everything is in view.
+
+        Parameters
+        ----------
+
+        data:
+            Collection of values that are being added to the figure.
+            Will be a list if coming from plot_ground_Truths or
+            plot_tracks, but will be a dictionary if coming from plot_measurements.
+        """
+
+        # fill in all data. If there is no data, fill all_x, all_y with current axis limits
+        if not data:
+            all_x = list(self.fig.layout.xaxis.range)
+            all_y = list(self.fig.layout.xaxis.range)
+        else:
+            all_x = list()
+            all_y = list()
+
+        # fill in data
+        if type == "measurements":
+
+            for key, item in data.items():
+                all_x.extend(data[key]["x"])
+                all_y.extend(data[key]["y"])
+
+        elif type in ("ground_truth", "tracks"):
+
+            for n, _ in enumerate(data):
+                all_x.extend(data[n]["x"])
+                all_y.extend(data[n]["y"])
+
+        elif type == "sensor":
+            sensor_xy = np.array([sensor.position[[0, 1], 0] for sensor in data])
+            all_x.extend(sensor_xy[:, 0])
+            all_y.extend(sensor_xy[:, 1])
+
+        elif type == "particle_or_uncertainty":
+            # data comes in format of list of dictionaries. Each dictionary contains 'x' and 'y',
+            # which are a list of lists.
+            for dictionary in data:
+                for x_values in dictionary["x"]:
+                    all_x.extend([np.nanmax(x_values), np.nanmin(x_values)])
+                for y_values in dictionary["y"]:
+                    all_y.extend([np.nanmax(y_values), np.nanmin(y_values)])
+
+        xmax = max(all_x)
+        ymax = max(all_y)
+        xmin = min(all_x)
+        ymin = min(all_y)
+
+        if self.equal_size:
+            xmax = ymax = max(xmax, ymax)
+            xmin = ymin = min(xmin, ymin)
+
+        # if it's first time plotting data, want to ensure plotter is bound to that data
+        # and not the default values. Issues arise if the initial plotted data is much
+        # smaller than the default 0 to 10 values.
+        if not self.plotting_function_called:
+
+            self.fig.update_xaxes(range=[xmin, xmax])
+            self.fig.update_yaxes(range=[ymin, ymax])
+
+        # need to check if it's actually necessary to resize or not
+        if xmax >= self.fig.layout.xaxis.range[1] or xmin <= self.fig.layout.xaxis.range[0]:
+
+            xmax = max(xmax, self.fig.layout.xaxis.range[1])
+            xmin = min(xmin, self.fig.layout.xaxis.range[0])
+            xrange = xmax - xmin
+
+            # update figure while adding a small buffer to the mins and maxes
+            self.fig.update_xaxes(range=[xmin - xrange / 20, xmax + xrange / 20])
+
+        if ymax >= self.fig.layout.yaxis.range[1] or ymin <= self.fig.layout.yaxis.range[0]:
+
+            ymax = max(ymax, self.fig.layout.yaxis.range[1])
+            ymin = min(ymin, self.fig.layout.yaxis.range[0])
+            yrange = ymax - ymin
+
+            self.fig.update_yaxes(range=[ymin - yrange / 20, ymax + yrange / 20])
+
+    def plot_ground_truths(self, truths, mapping, truths_label="Ground Truth",
+                           resize=True, **kwargs):
+
+        """Plots ground truth(s)
+
+        Plots each ground truth path passed in to :attr:`truths` and generates a legend
+        automatically. Ground truths are plotted as dashed lines with default colors.
+
+        Users can change linestyle, color and marker using keyword arguments. Any changes
+        will apply to all ground truths.
+
+        Parameters
+        ----------
+        truths : Collection of :class:`~.GroundTruthPath`
+            Collection of  ground truths which will be plotted. If not a collection and instead a
+            single :class:`~.GroundTruthPath` type, the argument is modified to be a set to allow
+            for iteration.
+        mapping: list
+            List of items specifying the mapping of the position components of the state space.
+        truths_label: str
+            Name of ground truths in legend/plot
+        resize: bool
+            if True, will resize figure to ensure that ground truths are in view
+        \\*\\*kwargs: dict
+            Additional arguments to be passed to plot function. Default is ``linestyle="--"``.
+
+        """
+
+        if not isinstance(truths, Collection) or isinstance(truths, StateMutableSequence):
+            truths = {truths}  # Make a set of length 1
+
+        data = [dict() for _ in truths]  # put all data into one place for later plotting
+        for n, truth in enumerate(truths):
+
+            # initialise arrays that go inside the dictionary
+            data[n].update(x=np.zeros(len(truth)),
+                           y=np.zeros(len(truth)),
+                           time=np.array([0 for _ in range(len(truth))], dtype=object),
+                           time_str=np.array([0 for _ in range(len(truth))], dtype=object),
+                           type=np.array([0 for _ in range(len(truth))], dtype=object))
+
+            for k, state in enumerate(truth):
+                # fill the arrays here
+                data[n]["x"][k] = state.state_vector[mapping[0]]
+                data[n]["y"][k] = state.state_vector[mapping[1]]
+                data[n]["time"][k] = state.timestamp
+                data[n]["time_str"][k] = str(state.timestamp)
+                data[n]["type"][k] = type(state).__name__
+
+        trace_base = len(self.fig.data)  # number of traces currently in the animation
+
+        # add a trace that keeps the legend up for the entire simulation (will remain
+        # even if no truths are present), then add a trace for each truth in the simulation.
+        # initialise keyword arguments, then add them to the traces
+        truth_kwargs = dict(x=[], y=[], mode="lines", hoverinfo='none', legendgroup=truths_label,
+                            line=dict(dash="dash", color=self.colorway[0]), legendrank=100,
+                            name=truths_label, showlegend=True)
+        truth_kwargs.update(kwargs)
+        # legend dummy trace
+        self.fig.add_trace(go.Scatter(truth_kwargs))
+
+        # we don't want the legend for any of the actual traces
+        truth_kwargs.update({"showlegend": False})
+
+        for n, _ in enumerate(truths):
+            # change the colour of each truth and include n in its name
+            truth_kwargs.update({
+                "line": dict(dash="dash", color=self.colorway[n % len(self.colorway)])})
+            truth_kwargs.update(kwargs)
+            self.fig.add_trace(go.Scatter(truth_kwargs))  # add to traces
+
+        for frame in self.fig.frames:
+
+            # get current fig data and traces
+            data_ = list(frame.data)
+            traces_ = list(frame.traces)
+
+            # convert string to datetime object
+            frame_time = datetime.fromisoformat(frame.name)
+            cutoff_time = (frame_time - self.time_window)
+
+            # for the legend
+            data_.append(go.Scatter(x=[0, 0], y=[0, 0]))
+            traces_.append(trace_base)
+
+            for n, truth in enumerate(truths):
+                # all truth points that come at or before the frame time
+                t_upper = [data[n]["time"] <= frame_time]
+
+                # only select detections that come after the time cut-off
+                t_lower = [data[n]["time"] >= cutoff_time]
+
+                # put together
+                mask = np.logical_and(t_upper, t_lower)
+
+                # find x, y, time, and type
+                truth_x = data[n]["x"][tuple(mask)]
+                # add in np.inf to ensure traces are present for every timestep
+                truth_x = np.append(truth_x, [np.inf])
+                truth_y = data[n]["y"][tuple(mask)]
+                truth_y = np.append(truth_y, [np.inf])
+                times = data[n]["time_str"][tuple(mask)]
+
+                data_.append(go.Scatter(x=truth_x,
+                                        y=truth_y,
+                                        meta=times,
+                                        hovertemplate='GroundTruthState' +
+                                                      '<br>(%{x}, %{y})' +
+                                                      '<br>Time: %{meta}'))
+
+                traces_.append(trace_base + n + 1)  # append data to correct trace
+
+                frame.data = data_
+                frame.traces = traces_
+
+        if resize:
+            self._resize(data, type="ground_truth")
+
+        # we have called a plotting function so update flag (gets used in _resize)
+        self.plotting_function_called = True
+
+    def plot_measurements(self, measurements, mapping, measurement_model=None,
+                          resize=True, measurements_label="Measurements",
+                          convert_measurements=True, **kwargs):
+        """Plots measurements
+
+        Plots detections and clutter, generating a legend automatically. Detections are plotted as
+        blue circles by default unless the detection type is clutter.
+        If the detection type is :class:`~.Clutter` it is plotted as a yellow 'tri-up' marker.
+
+        Users can change the color and marker of detections using keyword arguments but not for
+        clutter detections.
+
+        Parameters
+        ----------
+        measurements : Collection of :class:`~.Detection`
+            Detections which will be plotted. If measurements is a set of lists it is flattened.
+        mapping: list
+            List of items specifying the mapping of the position components of the state space.
+        measurement_model : :class:`~.Model`, optional
+            User-defined measurement model to be used in finding measurement state inverses if
+            they cannot be found from the measurements themselves.
+        resize: bool
+            If True, will resize figure to ensure measurements are in view
+        measurements_label : str
+            Label for the measurements.  Default is "Measurements".
+        convert_measurements : bool
+            Should the measurements be converted from measurement space to state space before
+            being plotted. Default is True
+        \\*\\*kwargs: dict
+            Additional arguments to be passed to scatter function for detections. Defaults are
+            ``marker=dict(color="#636EFA")``.
+        """
+
+        if not isinstance(measurements, Collection):
+            measurements = {measurements}  # Make a set of length 1
+
+        if any(isinstance(item, set) for item in measurements):
+            measurements_set = chain.from_iterable(measurements)  # Flatten into one set
+        else:
+            measurements_set = measurements
+        plot_detections, plot_clutter = self._conv_measurements(measurements_set,
+                                                                mapping,
+                                                                measurement_model,
+                                                                convert_measurements)
+        plot_combined = {'Detection': plot_detections,
+                         'Clutter': plot_clutter}  # for later reference
+
+        # this dictionary will store all the plotting data that we need
+        # from the detections and clutter into numpy arrays that we can easily
+        # access to plot
+        combined_data = dict()
+
+        # only add clutter or detections to plot if necessary
+        if plot_detections:
+            combined_data.update(dict(Detection=dict()))
+        if plot_clutter:
+            combined_data.update(dict(Clutter=dict()))
+
+        # initialise combined_data
+        for key in combined_data.keys():
+            length = len(plot_combined[key])
+            combined_data[key].update({
+                "x": np.zeros(length),
+                "y": np.zeros(length),
+                "time": np.array([0 for _ in range(length)], dtype=object),
+                "time_str": np.array([0 for _ in range(length)], dtype=object),
+                "type": np.array([0 for _ in range(length)], dtype=object)})
+
+        # and now fill in the data
+
+        for key in combined_data.keys():
+            for n, det in enumerate(plot_combined[key]):
+                x, y = list(plot_combined[key].values())[n]
+                combined_data[key]["x"][n] = x
+                combined_data[key]["y"][n] = y
+                combined_data[key]["time"][n] = det.timestamp
+                combined_data[key]["time_str"][n] = str(det.timestamp)
+                combined_data[key]["type"][n] = type(det).__name__
+
+        # get number of traces currently in fig
+        trace_base = len(self.fig.data)
+
+        # initialise detections
+        name = measurements_label + "<br>(Detections)"
+        measurement_kwargs = dict(x=[], y=[], mode='markers',
+                                  name=name,
+                                  legendgroup='Detections (Measurements)',
+                                  legendrank=200, showlegend=True,
+                                  marker=dict(color="#636EFA"), hoverinfo='none')
+        measurement_kwargs.update(kwargs)
+
+        self.fig.add_trace(go.Scatter(measurement_kwargs))  # trace for legend
+
+        measurement_kwargs.update({"showlegend": False})
+        self.fig.add_trace(go.Scatter(measurement_kwargs))  # trace for plotting
+
+        # change necessary kwargs to initialise clutter trace
+        name = measurements_label + "<br>(Clutter)"
+        measurement_kwargs.update({"legendgroup": 'Clutter', "legendrank": 300,
+                                   "marker": dict(symbol="star-triangle-up", color='#FECB52'),
+                                   "name": name, 'showlegend': True})
+
+        self.fig.add_trace(go.Scatter(measurement_kwargs))  # trace for plotting clutter
+
+        # add data to frames
+        for frame in self.fig.frames:
+
+            data_ = list(frame.data)
+            traces_ = list(frame.traces)
+
+            # add blank data to ensure detection legend stays in place
+            data_.append(go.Scatter(x=[-np.inf, np.inf], y=[-np.inf, np.inf]))
+            traces_.append(trace_base)  # ensure data is added to correct trace
+
+            frame_time = datetime.fromisoformat(frame.name)  # convert string to datetime object
+
+            # time at which dets will disappear from the fig
+            cutoff_time = (frame_time - self.time_window)
+
+            for j, key in enumerate(combined_data.keys()):
+                # only select measurements that arrive by the time of the current frame
+                t_upper = [combined_data[key]["time"] <= frame_time]
+
+                # only select detections that come after the time cut-off
+                t_lower = [combined_data[key]["time"] >= cutoff_time]
+
+                # put them together to create the final mask
+                mask = np.logical_and(t_upper, t_lower)
+
+                # find x and y points for true detections and clutter
+                det_x = combined_data[key]["x"][tuple(mask)]
+                det_x = np.append(det_x, [np.inf])
+                det_y = combined_data[key]["y"][tuple(mask)]
+                det_y = np.append(det_y, [np.inf])
+                det_times = combined_data[key]["time_str"][tuple(mask)]
+
+                data_.append(go.Scatter(x=det_x,
+                                        y=det_y,
+                                        meta=det_times,
+                                        hovertemplate=f'{key}' +
+                                                      '<br>(%{x}, %{y})' +
+                                                      '<br>Time: %{meta}'))
+                traces_.append(trace_base + j + 1)
+
+            frame.data = data_  # update the figure
+            frame.traces = traces_
+
+        if resize:
+            self._resize(combined_data, "measurements")
+
+        # we have called a plotting function so update flag (gets used in resize)
+        self.plotting_function_called = True
+
+    def plot_tracks(self, tracks, mapping, uncertainty=False, resize=True,
+                    particle=False, plot_history=False, ellipse_points=30,
+                    track_label="Tracks", **kwargs):
+        """
+        Plots each track generated, generating a legend automatically. If 'uncertainty=True',
+        error ellipses are plotted. Tracks are plotted as solid lines with point markers
+        and default colours.
+
+        Users can change linestyle, color, and marker using keyword arguments. Uncertainty metrics
+        will also be plotted with the user defined colour and any changes will apply to all tracks.
+
+        Parameters
+        ----------
+        tracks: Collection of :class '~Track'
+            Collection of tracks which will be plotted. If not a collection, and instead a single
+            :class:'~Track' type, the argument is modified to be a set to allow for iteration
+
+        mapping: list
+            List of items specifying the mapping of the position
+            components of the state space
+        uncertainty: bool
+            If True, function plots uncertainty ellipses
+        resize: bool
+            If True, plotter will change bounds so that tracks are in view
+        particle: bool
+            If True, function plots particles
+        plot_history: bool
+            If true, plots all particles and uncertainty ellipses up to current time step
+        ellipse_points: int
+            Number of points for polygon approximating ellipse shape
+        track_label: str
+            Label to apply to all tracks for legend
+        \\*\\*kwargs: dict
+            Additional arguments to be passed to plot function. Defaults are ``linestyle="-"``,
+            ``marker='s'`` for :class:`~.Update` and ``marker='o'`` for other states.
+
+        Returns
+        -------
+        """
+
+        if not isinstance(tracks, Collection) or isinstance(tracks, StateMutableSequence):
+            tracks = {tracks}  # Make a set of length 1
+
+        # So that we can plot tracks for both the current time and for some previous times,
+        # we put plotting data for each track into a dictionary so that it can be easily
+        # accessed later.
+        data = [dict() for _ in tracks]
+
+        for n, track in enumerate(tracks):  # sum up means - accounts for particle filter
+
+            xydata = np.concatenate(
+                [(getattr(state, 'mean', state.state_vector)[mapping, :])
+                 for state in track],
+                axis=1)
+
+            # initialise arrays that go inside the dictionary
+            data[n].update(x=xydata[0],
+                           y=xydata[1],
+                           time=np.array([0 for _ in range(len(track))], dtype=object),
+                           time_str=np.array([0 for _ in range(len(track))], dtype=object),
+                           type=np.array([0 for _ in range(len(track))], dtype=object))
+
+            for k, state in enumerate(track):
+                # fill the arrays here
+                data[n]["time"][k] = state.timestamp
+                data[n]["time_str"][k] = str(state.timestamp)
+                data[n]["type"][k] = type(state).__name__
+
+        trace_base = len(self.fig.data)  # number of traces
+
+        # add dummy trace for legend for track
+
+        track_kwargs = dict(x=[], y=[], mode="markers+lines", line=dict(color=self.colorway[2]),
+                            legendgroup=track_label, legendrank=400, name=track_label,
+                            showlegend=True)
+        track_kwargs.update(kwargs)
+        self.fig.add_trace(go.Scatter(track_kwargs))
+
+        # and initialise traces for every track. Need to change a few kwargs:
+        track_kwargs.update({'showlegend': False})
+
+        for k, _ in enumerate(tracks):
+            # update track colours
+            track_kwargs.update({'line': dict(color=self.colorway[(k + 2) % len(self.colorway)])})
+            track_kwargs.update(kwargs)
+            self.fig.add_trace(go.Scatter(track_kwargs))
+
+        for frame in self.fig.frames:
+            # get current fig data and traces
+            data_ = list(frame.data)
+            traces_ = list(frame.traces)
+
+            # convert string to datetime object
+            frame_time = datetime.fromisoformat(frame.name)
+
+            self.all_masks[frame_time] = dict()  # save mask for later use
+            cutoff_time = (frame_time - self.time_window)
+            # add blank data to ensure legend stays in place
+            data_.append(go.Scatter(x=[-np.inf, np.inf], y=[-np.inf, np.inf]))
+            traces_.append(trace_base)  # ensure data is added to correct trace
+
+            for n, track in enumerate(tracks):
+
+                # all track points that come at or before the frame time
+                t_upper = [data[n]["time"] <= frame_time]
+                # only select detections that come after the time cut-off
+                t_lower = [data[n]["time"] >= cutoff_time]
+
+                # put together
+                mask = np.logical_and(t_upper, t_lower)
+
+                # put into dictionary for later use
+                if plot_history:
+                    self.all_masks[frame_time][n] = np.logical_and(t_upper, t_lower)
+                else:
+                    self.all_masks[frame_time][n] = [data[n]["time"] == frame_time]
+
+                # find x, y, time, and type
+                track_x = data[n]["x"][tuple(mask)]
+                # add np.inf to plot so that the traces are present for entire simulation
+                track_x = np.append(track_x, [np.inf])
+
+                # repeat for y
+                track_y = data[n]["y"][tuple(mask)]
+                track_y = np.append(track_y, [np.inf])
+                track_type = data[n]["type"][tuple(mask)]
+                times = data[n]["time_str"][tuple(mask)]
+
+                data_.append(go.Scatter(x=track_x,  # plot track
+                                        y=track_y,
+                                        meta=track_type,
+                                        customdata=times,
+                                        hovertemplate='%{meta}' +
+                                                      '<br>(%{x}, %{y})' +
+                                                      '<br>Time: %{customdata}'))
+
+                traces_.append(trace_base + n + 1)  # add to correct trace
+
+                frame.data = data_
+                frame.traces = traces_
+
+        if resize:
+            self._resize(data, "tracks")
+
+        if uncertainty:  # plot ellipses
+
+            uncertainty_kwargs = dict(x=[], y=[], legendgroup='Uncertainty', fill='toself',
+                                      fillcolor=self.colorway[2],
+                                      opacity=0.2, legendrank=500, name='Track<br>Uncertainty',
+                                      hoverinfo='skip',
+                                      mode='none', showlegend=True)
+            uncertainty_kwargs.update(kwargs)
+
+            # dummy trace for legend for uncertainty
+            self.fig.add_trace(go.Scatter(uncertainty_kwargs))
+
+            # and an uncertainty ellipse trace for each track
+            uncertainty_kwargs.update({'showlegend': False})
+            for k, _ in enumerate(tracks):
+                uncertainty_kwargs.update(
+                    {'fillcolor': self.colorway[(k + 2) % len(self.colorway)]})
+                uncertainty_kwargs.update(kwargs)
+                self.fig.add_trace(go.Scatter(uncertainty_kwargs))
+
+            # following function finds uncertainty data points and plots them
+            self._plot_particles_and_ellipses(tracks, mapping, resize, method="uncertainty")
+
+        if particle:  # plot particles
+
+            # initialise traces. One for legend and one per track
+
+            particle_kwargs = dict(mode='markers', marker=dict(size=2, color=self.colorway[2]),
+                                   opacity=0.4,
+                                   hoverinfo='skip', legendgroup='particles', name='particles',
+                                   legendrank=520, showlegend=True)
+            # apply any keyword arguments
+            particle_kwargs.update(kwargs)
+            self.fig.add_trace(go.Scatter(particle_kwargs))  # legend trace
+
+            particle_kwargs.update({"showlegend": False})
+
+            for k, track in enumerate(tracks):  # trace for each track
+
+                particle_kwargs.update(
+                    {'marker': dict(size=2, color=self.colorway[(k + 2) % len(self.colorway)])})
+                particle_kwargs.update(kwargs)
+                self.fig.add_trace(go.Scatter(particle_kwargs))
+
+            self._plot_particles_and_ellipses(tracks, mapping, resize, method="particles")
+
+        # we have called a plotting function so update flag
+        self.plotting_function_called = True
+
+    def _plot_particles_and_ellipses(self, tracks, mapping, resize, method="uncertainty"):
+
+        """
+        The logic for plotting uncertainty ellipses and particles is nearly identical,
+        so it is put into one function.
+
+        Parameters
+        ----------
+        tracks: Collection of :class '~Track'
+            Collection of tracks which will be plotted. If not a collection, and instead a single
+            :class:'~Track' type, the argument is modified to be a set to allow for iteration
+        mapping: list
+            List of items specifying the mapping of the position components of the state space.
+        method: str
+            Can either be "uncertainty" or "particles". Depends on what the function is plotting.
+        """
+
+        data = [dict() for _ in tracks]
+        trace_base = len(self.fig.data)
+        for n, track in enumerate(tracks):
+
+            # initialise arrays that store particle/ellipse for later plotting
+            data[n].update(x=np.array([0 for _ in range(len(track))], dtype=object),
+                           y=np.array([0 for _ in range(len(track))], dtype=object))
+
+            for k, state in enumerate(track):
+
+                # find data points
+                if method == "uncertainty":
+
+                    data_x, data_y = Plotterly._generate_ellipse_points(state, mapping)
+                    data_x = list(data_x)
+                    data_y = list(data_y)
+                    data_x.append(np.nan)  # necessary to draw multiple ellipses at once
+                    data_y.append(np.nan)
+                    data[n]["x"][k] = data_x
+                    data[n]["y"][k] = data_y
+
+                elif method == "particles":
+
+                    data_xy = state.state_vector[mapping[:2], :]
+                    data[n]["x"][k] = data_xy[0]
+                    data[n]["y"][k] = data_xy[1]
+
+                else:
+                    raise ValueError("Should be 'uncertainty' or 'particles'")
+
+        for frame in self.fig.frames:
+
+            frame_time = datetime.fromisoformat(frame.name)
+
+            data_ = list(frame.data)  # current data in frame
+            traces_ = list(frame.traces)  # current traces in frame
+
+            data_.append(go.Scatter(x=[-np.inf], y=[np.inf]))  # add empty data for legend trace
+            traces_.append(trace_base - len(tracks) - 1)  # ensure correct trace
+
+            for n, track in enumerate(tracks):
+                # now plot the data
+                _x = list(chain(*data[n]["x"][tuple(self.all_masks[frame_time][n])]))
+                _y = list(chain(*data[n]["y"][tuple(self.all_masks[frame_time][n])]))
+                _x.append(np.inf)
+                _y.append(np.inf)
+                data_.append(go.Scatter(x=_x, y=_y))
+                traces_.append(trace_base - len(tracks) + n)
+
+            frame.data = data_
+            frame.traces = traces_
+
+        if resize:
+            self._resize(data, type="particle_or_uncertainty")
+
+    def plot_sensors(self, sensors, sensor_label="Sensors", resize=True, **kwargs):
+        """Plots sensor(s)
+
+        Plots sensors.  Users can change the color and marker of detections using keyword
+        arguments. Default is a black 'x' marker. Currently only works for stationary
+        sensors.
+
+        Parameters
+        ----------
+        sensors : Collection of :class:`~.Sensor`
+            Sensors to plot
+        sensor_label: str
+            Label to apply to all tracks for legend.
+        \\*\\*kwargs: dict
+            Additional arguments to be passed to scatter function for detections. Defaults are
+            ``marker=dict(symbol='x', color='black')``.
+        """
+        if not isinstance(sensors, Collection):
+            sensors = {sensors}
+
+        # don't run any of this if there is no data input
+        if sensors:
+            trace_base = len(self.fig.data)  # number of traces currently in figure
+            sensor_kwargs = dict(mode='markers', marker=dict(symbol='x', color='black'),
+                                 legendgroup=sensor_label, legendrank=50,
+                                 name=sensor_label, showlegend=True)
+            sensor_kwargs.update(kwargs)
+
+            self.fig.add_trace(go.Scatter(sensor_kwargs))  # initialises trace
+
+            # sensor position
+            sensor_xy = np.array([sensor.position[[0, 1], 0] for sensor in sensors])
+            if resize:
+                self._resize(sensors, "sensor")
+
+            for frame in self.fig.frames:  # the plotting bit
+                traces_ = list(frame.traces)
+                data_ = list(frame.data)
+
+                data_.append(go.Scatter(x=sensor_xy[:, 0], y=sensor_xy[:, 1]))
+                traces_.append(trace_base)
+
+                frame.traces = traces_
+                frame.data = data_
+
+        # we have called a plotting function so update flag (used in _resize)
+        self.plotting_function_called = True
