@@ -1,9 +1,10 @@
 import datetime
-from typing import Set
+from typing import Set, Union
+from itertools import combinations
 
 from ..base import Property
 from .base import Type
-from .time import TimeRange
+from .time import TimeRange, CompoundTimeRange
 
 
 class Association(Type):
@@ -48,8 +49,12 @@ class TimeRangeAssociation(Association):
     range of times
     """
 
-    time_range: TimeRange = Property(
+    time_range: Union[CompoundTimeRange, TimeRange] = Property(
         default=None, doc="Range of times that association exists over. Default is None")
+
+    @property
+    def duration(self):
+        return self.time_range.duration.total_seconds()
 
 
 class AssociationSet(Type):
@@ -66,6 +71,92 @@ class AssociationSet(Type):
         super().__init__(associations, *args, **kwargs)
         if self.associations is None:
             self.associations = set()
+        if not all(isinstance(member, Association) for member in self.associations):
+            raise TypeError("Association set must contain only Association instances")
+        self._simplify()
+
+    def __eq__(self, other):
+        return self.associations == other.associations
+
+    def add(self, association):
+        if association is None:
+            return
+        elif isinstance(association, Association):
+            self.associations.add(association)
+        elif isinstance(association, AssociationSet):
+            for component in association:
+                self.add(component)
+        else:
+            raise TypeError("Supplied parameter must be an Association or AssociationSet")
+        self._simplify()
+
+    def _simplify(self):
+        """Where multiple associations describe the same pair of objects, combine them into one.
+        This is only implemented for pairs with a time_range attribute - others will be skipped
+        """
+        to_remove = set()
+        for (assoc1, assoc2) in combinations(self.associations, 2):
+            if not (len(assoc1.objects) == 2 and len(assoc2.objects) == 2) or \
+                    not(hasattr(assoc1, 'time_range') and hasattr(assoc2, 'time_range')):
+                continue
+            if assoc1.objects == assoc2.objects:
+                if isinstance(assoc1.time_range, CompoundTimeRange):
+                    assoc1.time_range.add(assoc2.time_range)
+                    to_remove.add(assoc2)
+                elif isinstance(assoc2.time_range, CompoundTimeRange):
+                    assoc2.time_range.add(assoc1.time_range)
+                    to_remove.add(assoc1)
+                else:
+                    assoc1.time_range = CompoundTimeRange([assoc1.time_range, assoc2.time_range])
+                    to_remove.add(assoc2)
+        for assoc in to_remove:
+            self.remove(assoc)
+
+    def remove(self, association):
+        if association is None:
+            return
+        elif isinstance(association, Association):
+            if association not in self.associations:
+                raise ValueError("Supplied parameter must be contained by this instance")
+            self.associations.remove(association)
+        elif isinstance(association, AssociationSet):
+            for component in association:
+                self.remove(component)
+        else:
+            raise TypeError("Supplied parameter must be an Association or AssociationSet")
+
+    @property
+    def key_times(self):
+        """Returns all timestamps at which a component starts or ends, or where there is a
+        :class:`.~SingleTimeAssociation`."""
+        key_times = list(self.overall_time_range.key_times)
+        for association in self.associations:
+            if isinstance(association, SingleTimeAssociation):
+                key_times.append(association.timestamp)
+        return sorted(key_times)
+
+    @property
+    def overall_time_range(self):
+        """Returns a :class:`~.CompoundTimeRange` covering all times at which at least
+        one association is active.
+
+        Note: :class:`~.SingleTimeAssociation` are not counted
+        """
+        overall_range = CompoundTimeRange()
+        for association in self.associations:
+            if hasattr(association, 'time_range'):
+                overall_range.add(association.time_range)
+        return overall_range
+
+    @property
+    def object_set(self):
+        """Returns a set of all objects contained by this instance.
+        """
+        object_set = set()
+        for assoc in self.associations:
+            for obj in assoc.objects:
+                object_set.add(obj)
+        return object_set
 
     def associations_at_timestamp(self, timestamp):
         """Return the associations that exist at a given timestamp
@@ -80,9 +171,11 @@ class AssociationSet(Type):
 
         Returns
         -------
-        : set of :class:`~.Association`
+        : :class:`~.AssociationSet`
             Associations which occur at specified timestamp
         """
+        if not isinstance(timestamp, datetime.datetime):
+            raise TypeError("Supplied parameter must be a datetime.datetime object")
         ret_associations = set()
         for association in self.associations:
             # If the association is at a single time
@@ -92,7 +185,7 @@ class AssociationSet(Type):
             else:
                 if timestamp in association.time_range:
                     ret_associations.add(association)
-        return ret_associations
+        return AssociationSet(ret_associations)
 
     def associations_including_objects(self, objects):
         """Return associations that include all the given objects
@@ -106,18 +199,17 @@ class AssociationSet(Type):
             Set of objects to look for in associations
         Returns
         -------
-        : set of :class:`~.Association`
-            A set of objects which have been associated
+        : class:`~.AssociationSet`
+            A set of associations containing every member of objects
         """
-
         # Ensure objects is iterable
         if not isinstance(objects, list) and not isinstance(objects, set):
             objects = {objects}
 
-        return {association
-                for association in self.associations
-                for object_ in objects
-                if object_ in association.objects}
+        return AssociationSet({association
+                              for association in self.associations
+                              if all(object_ in association.objects
+                                     for object_ in objects)})
 
     def __contains__(self, item):
         return item in self.associations
