@@ -1,12 +1,19 @@
-from collections.abc import Sized, Iterable, Container
+import copy
+from collections import abc
 from typing import MutableSequence
 
-from .base import Type
+import numpy as np
+
 from ..base import Property
-from .state import TaggedWeightedGaussianState, WeightedGaussianState
+from ..functions import gm_reduce_single
+from .base import Type
+from .array import StateVectors
+from .numeric import Probability
+from .prediction import Prediction, GaussianStatePrediction
+from .state import GaussianState, TaggedWeightedGaussianState, WeightedGaussianState
 
 
-class GaussianMixture(Type, Sized, Iterable, Container):
+class GaussianMixture(Type, abc.MutableSequence):
     """
     Gaussian Mixture type
 
@@ -28,6 +35,8 @@ class GaussianMixture(Type, Sized, Iterable, Container):
                 for component in self.components):
             raise ValueError("Cannot form GaussianMixtureState out of "
                              "non-WeightedGaussianState inputs!")
+        if len({component.timestamp for component in self.components}) > 1:
+            raise ValueError("All components must have the same timestamp")
 
     def __contains__(self, index):
         # check if 'components' contains any WeightedGaussianState
@@ -37,24 +46,73 @@ class GaussianMixture(Type, Sized, Iterable, Container):
         else:
             raise ValueError("Index must be WeightedGaussianState")
 
-    def __iter__(self):
-        return iter(self.components)
-
     def __getitem__(self, index):
         # retrieve WeightedGaussianState by array index
         return self.components[index]
 
     def __setitem__(self, index, value):
+        if not isinstance(value, (WeightedGaussianState, TaggedWeightedGaussianState)):
+            raise ValueError("Cannot form GaussianMixtureState out of "
+                             "non-WeightedGaussianState inputs!")
         return self.components.__setitem__(index, value)
+
+    def __delitem__(self, value):
+        return self.components.__delitem__(value)
 
     def __len__(self):
         return len(self.components)
 
-    def append(self, component):
-        return self.components.append(component)
+    def __copy__(self):
+        inst = self.__class__.__new__(self.__class__)
+        inst.__dict__.update(self.__dict__)
+        property_name = self.__class__.components._property_name
+        inst.__dict__[property_name] = copy.copy(self.__dict__[property_name])
+        return inst
 
-    def extend(self, new_components):
-        return self.components.extend(new_components)
+    def insert(self, index, value):
+        if not isinstance(value, (WeightedGaussianState, TaggedWeightedGaussianState)):
+            raise ValueError("Cannot form GaussianMixtureState out of "
+                             "non-WeightedGaussianState inputs!")
+        return self.components.insert(index, value)
+
+    @property
+    def ndim(self):
+        if len(self.components):
+            return self.components[0].ndim
+        return 0
+
+    @property
+    def means(self):
+        return StateVectors([component.mean for component in self.components])
+
+    @property
+    def covars(self):
+        return np.stack([component.covar for component in self.components], axis=2)
+
+    @property
+    def weights(self):
+        return np.asarray([component.weight for component in self.components])
+
+    @property
+    def state_vector(self):
+        return self.mean
+
+    @property
+    def mean(self):
+        means = self.means
+        weights = self.weights / Probability.sum(self.weights)
+        return np.average(means, axis=1, weights=weights)
+
+    @property
+    def covar(self):
+        _, covar = gm_reduce_single(self.means, self.covars,
+                                    self.weights)
+        return covar
+
+    @property
+    def timestamp(self):
+        """Timestamp"""
+        return next((component.timestamp for component in self.components), None)
 
     @property
     def component_tags(self):
@@ -67,3 +125,7 @@ class GaussianMixture(Type, Sized, Iterable, Container):
             raise ValueError("All components must be "
                              "TaggedWeightedGaussianState!")
         return component_tags
+
+
+GaussianState.register(GaussianMixture)
+Prediction.class_mapping[Prediction][GaussianMixture] = GaussianStatePrediction
